+++ packages/python_ffi_cpython_dart/lib/src/ffi/generated_bindings.g.dart
@@ -47,8 +47,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyMem_CallocPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Pointer<ffi.Void> Function(ffi.Size, ffi.Size)>>('PyMem_Calloc');
+          ffi
+          .NativeFunction<ffi.Pointer<ffi.Void> Function(ffi.Size, ffi.Size)>>(
+      'PyMem_Calloc');
   late final _PyMem_Calloc =
       _PyMem_CallocPtr.asFunction<ffi.Pointer<ffi.Void> Function(int, int)>();
 
@@ -108,9 +109,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyMem_RawCallocPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Pointer<ffi.Void> Function(
-              ffi.Size, ffi.Size)>>('PyMem_RawCalloc');
+          ffi
+          .NativeFunction<ffi.Pointer<ffi.Void> Function(ffi.Size, ffi.Size)>>(
+      'PyMem_RawCalloc');
   late final _PyMem_RawCalloc = _PyMem_RawCallocPtr.asFunction<
       ffi.Pointer<ffi.Void> Function(int, int)>();
 
@@ -379,9 +380,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyBuffer_IsContiguousPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<Py_buffer>, ffi.Char)>>('PyBuffer_IsContiguous');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<Py_buffer>, ffi.Char)>>(
+      'PyBuffer_IsContiguous');
   late final _PyBuffer_IsContiguous = _PyBuffer_IsContiguousPtr.asFunction<
       int Function(ffi.Pointer<Py_buffer>, int)>();
 
@@ -1880,9 +1881,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyTrash_condPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<PyObject>, destructor)>>('_PyTrash_cond');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, destructor)>>(
+      '_PyTrash_cond');
   late final __PyTrash_cond = __PyTrash_condPtr
       .asFunction<int Function(ffi.Pointer<PyObject>, destructor)>();
 
@@ -1911,9 +1912,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyObject_CallocPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Pointer<ffi.Void> Function(
-              ffi.Size, ffi.Size)>>('PyObject_Calloc');
+          ffi
+          .NativeFunction<ffi.Pointer<ffi.Void> Function(ffi.Size, ffi.Size)>>(
+      'PyObject_Calloc');
   late final _PyObject_Calloc = _PyObject_CallocPtr.asFunction<
       ffi.Pointer<ffi.Void> Function(int, int)>();
 
@@ -2544,9 +2545,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyByteArray_ResizePtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<PyObject>, Py_ssize_t)>>('PyByteArray_Resize');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyByteArray_Resize');
   late final _PyByteArray_Resize = _PyByteArray_ResizePtr.asFunction<
       int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -3146,9 +3147,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicode_ReadCharPtr = _lookup<
-      ffi.NativeFunction<
-          Py_UCS4 Function(
-              ffi.Pointer<PyObject>, Py_ssize_t)>>('PyUnicode_ReadChar');
+          ffi
+          .NativeFunction<Py_UCS4 Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicode_ReadChar');
   late final _PyUnicode_ReadChar = _PyUnicode_ReadCharPtr.asFunction<
       int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -4571,9 +4572,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicode_NewPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Pointer<PyObject> Function(
-              Py_ssize_t, Py_UCS4)>>('PyUnicode_New');
+          ffi
+          .NativeFunction<ffi.Pointer<PyObject> Function(Py_ssize_t, Py_UCS4)>>(
+      'PyUnicode_New');
   late final _PyUnicode_New =
       _PyUnicode_NewPtr.asFunction<ffi.Pointer<PyObject> Function(int, int)>();
 
@@ -6809,9 +6810,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyFloat_Unpack2Ptr = _lookup<
-      ffi.NativeFunction<
-          ffi.Double Function(
-              ffi.Pointer<ffi.Char>, ffi.Int)>>('PyFloat_Unpack2');
+          ffi
+          .NativeFunction<ffi.Double Function(ffi.Pointer<ffi.Char>, ffi.Int)>>(
+      'PyFloat_Unpack2');
   late final _PyFloat_Unpack2 = _PyFloat_Unpack2Ptr.asFunction<
       double Function(ffi.Pointer<ffi.Char>, int)>();
 
@@ -6826,9 +6827,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyFloat_Unpack4Ptr = _lookup<
-      ffi.NativeFunction<
-          ffi.Double Function(
-              ffi.Pointer<ffi.Char>, ffi.Int)>>('PyFloat_Unpack4');
+          ffi
+          .NativeFunction<ffi.Double Function(ffi.Pointer<ffi.Char>, ffi.Int)>>(
+      'PyFloat_Unpack4');
   late final _PyFloat_Unpack4 = _PyFloat_Unpack4Ptr.asFunction<
       double Function(ffi.Pointer<ffi.Char>, int)>();
 
@@ -6843,9 +6844,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyFloat_Unpack8Ptr = _lookup<
-      ffi.NativeFunction<
-          ffi.Double Function(
-              ffi.Pointer<ffi.Char>, ffi.Int)>>('PyFloat_Unpack8');
+          ffi
+          .NativeFunction<ffi.Double Function(ffi.Pointer<ffi.Char>, ffi.Int)>>(
+      'PyFloat_Unpack8');
   late final _PyFloat_Unpack8 = _PyFloat_Unpack8Ptr.asFunction<
       double Function(ffi.Pointer<ffi.Char>, int)>();
 
@@ -11014,9 +11015,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyInterpreterState_ClearPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(
-              ffi.Pointer<PyInterpreterState>)>>('PyInterpreterState_Clear');
+          ffi
+          .NativeFunction<ffi.Void Function(ffi.Pointer<PyInterpreterState>)>>(
+      'PyInterpreterState_Clear');
   late final _PyInterpreterState_Clear = _PyInterpreterState_ClearPtr
       .asFunction<void Function(ffi.Pointer<PyInterpreterState>)>();
 
@@ -11029,9 +11030,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyInterpreterState_DeletePtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(
-              ffi.Pointer<PyInterpreterState>)>>('PyInterpreterState_Delete');
+          ffi
+          .NativeFunction<ffi.Void Function(ffi.Pointer<PyInterpreterState>)>>(
+      'PyInterpreterState_Delete');
   late final _PyInterpreterState_Delete = _PyInterpreterState_DeletePtr
       .asFunction<void Function(ffi.Pointer<PyInterpreterState>)>();
 
@@ -11070,9 +11071,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyInterpreterState_GetIDPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int64 Function(
-              ffi.Pointer<PyInterpreterState>)>>('PyInterpreterState_GetID');
+          ffi
+          .NativeFunction<ffi.Int64 Function(ffi.Pointer<PyInterpreterState>)>>(
+      'PyInterpreterState_GetID');
   late final _PyInterpreterState_GetID = _PyInterpreterState_GetIDPtr
       .asFunction<int Function(ffi.Pointer<PyInterpreterState>)>();
 
@@ -11305,8 +11306,8 @@ base class DartPythonCBindings {
   }
 
   late final __PyInterpreterState_RequiresIDRefPtr = _lookup<
-          ffi.NativeFunction<
-              ffi.Int Function(ffi.Pointer<PyInterpreterState>)>>(
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyInterpreterState>)>>(
       '_PyInterpreterState_RequiresIDRef');
   late final __PyInterpreterState_RequiresIDRef =
       __PyInterpreterState_RequiresIDRefPtr
@@ -11719,9 +11720,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyCrossInterpreterData_LookupPtr = _lookup<
-      ffi.NativeFunction<
-          crossinterpdatafunc Function(
-              ffi.Pointer<PyObject>)>>('_PyCrossInterpreterData_Lookup');
+          ffi
+          .NativeFunction<crossinterpdatafunc Function(ffi.Pointer<PyObject>)>>(
+      '_PyCrossInterpreterData_Lookup');
   late final __PyCrossInterpreterData_Lookup =
       __PyCrossInterpreterData_LookupPtr
           .asFunction<crossinterpdatafunc Function(ffi.Pointer<PyObject>)>();
@@ -12330,9 +12331,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyWeakref_GetWeakrefCountPtr = _lookup<
-      ffi.NativeFunction<
-          Py_ssize_t Function(
-              ffi.Pointer<PyWeakReference>)>>('_PyWeakref_GetWeakrefCount');
+          ffi
+          .NativeFunction<Py_ssize_t Function(ffi.Pointer<PyWeakReference>)>>(
+      '_PyWeakref_GetWeakrefCount');
   late final __PyWeakref_GetWeakrefCount = __PyWeakref_GetWeakrefCountPtr
       .asFunction<int Function(ffi.Pointer<PyWeakReference>)>();
 
@@ -12873,9 +12874,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyTime_AsTimespecPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              _PyTime_t, ffi.Pointer<timespec>)>>('_PyTime_AsTimespec');
+          ffi
+          .NativeFunction<ffi.Int Function(_PyTime_t, ffi.Pointer<timespec>)>>(
+      '_PyTime_AsTimespec');
   late final __PyTime_AsTimespec = __PyTime_AsTimespecPtr
       .asFunction<int Function(int, ffi.Pointer<timespec>)>();
 
@@ -12890,9 +12891,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyTime_AsTimespec_clampPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(
-              _PyTime_t, ffi.Pointer<timespec>)>>('_PyTime_AsTimespec_clamp');
+          ffi
+          .NativeFunction<ffi.Void Function(_PyTime_t, ffi.Pointer<timespec>)>>(
+      '_PyTime_AsTimespec_clamp');
   late final __PyTime_AsTimespec_clamp = __PyTime_AsTimespec_clampPtr
       .asFunction<void Function(int, ffi.Pointer<timespec>)>();
 
@@ -12925,9 +12926,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyTime_MulDivPtr = _lookup<
-      ffi.NativeFunction<
-          _PyTime_t Function(
-              _PyTime_t, _PyTime_t, _PyTime_t)>>('_PyTime_MulDiv');
+          ffi
+          .NativeFunction<_PyTime_t Function(_PyTime_t, _PyTime_t, _PyTime_t)>>(
+      '_PyTime_MulDiv');
   late final __PyTime_MulDiv =
       __PyTime_MulDivPtr.asFunction<int Function(int, int, int)>();
 
@@ -14678,9 +14679,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyErr_BadInternalCallPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(
-              ffi.Pointer<ffi.Char>, ffi.Int)>>('_PyErr_BadInternalCall');
+          ffi
+          .NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Char>, ffi.Int)>>(
+      '_PyErr_BadInternalCall');
   late final __PyErr_BadInternalCall = __PyErr_BadInternalCallPtr
       .asFunction<void Function(ffi.Pointer<ffi.Char>, int)>();
 
@@ -14792,9 +14793,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyErr_SyntaxLocationPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Void Function(
-              ffi.Pointer<ffi.Char>, ffi.Int)>>('PyErr_SyntaxLocation');
+          ffi
+          .NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Char>, ffi.Int)>>(
+      'PyErr_SyntaxLocation');
   late final _PyErr_SyntaxLocation = _PyErr_SyntaxLocationPtr.asFunction<
       void Function(ffi.Pointer<ffi.Char>, int)>();
 
@@ -15011,9 +15012,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeEncodeError_SetStartPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeEncodeError_SetStart');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeEncodeError_SetStart');
   late final _PyUnicodeEncodeError_SetStart = _PyUnicodeEncodeError_SetStartPtr
       .asFunction<int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -15028,9 +15029,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeDecodeError_SetStartPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeDecodeError_SetStart');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeDecodeError_SetStart');
   late final _PyUnicodeDecodeError_SetStart = _PyUnicodeDecodeError_SetStartPtr
       .asFunction<int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -15045,9 +15046,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeTranslateError_SetStartPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeTranslateError_SetStart');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeTranslateError_SetStart');
   late final _PyUnicodeTranslateError_SetStart =
       _PyUnicodeTranslateError_SetStartPtr.asFunction<
           int Function(ffi.Pointer<PyObject>, int)>();
@@ -15117,9 +15118,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeEncodeError_SetEndPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeEncodeError_SetEnd');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeEncodeError_SetEnd');
   late final _PyUnicodeEncodeError_SetEnd = _PyUnicodeEncodeError_SetEndPtr
       .asFunction<int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -15134,9 +15135,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeDecodeError_SetEndPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeDecodeError_SetEnd');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeDecodeError_SetEnd');
   late final _PyUnicodeDecodeError_SetEnd = _PyUnicodeDecodeError_SetEndPtr
       .asFunction<int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -15151,9 +15152,9 @@ base class DartPythonCBindings {
   }
 
   late final _PyUnicodeTranslateError_SetEndPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(ffi.Pointer<PyObject>,
-              Py_ssize_t)>>('PyUnicodeTranslateError_SetEnd');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PyUnicodeTranslateError_SetEnd');
   late final _PyUnicodeTranslateError_SetEnd =
       _PyUnicodeTranslateError_SetEndPtr.asFunction<
           int Function(ffi.Pointer<PyObject>, int)>();
@@ -15982,9 +15983,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyThread_at_fork_reinitPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<PyThread_type_lock>)>>('_PyThread_at_fork_reinit');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyThread_type_lock>)>>(
+      '_PyThread_at_fork_reinit');
   late final __PyThread_at_fork_reinit = __PyThread_at_fork_reinitPtr
       .asFunction<int Function(ffi.Pointer<PyThread_type_lock>)>();
 
@@ -18292,9 +18293,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyOS_URandomPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<ffi.Void>, Py_ssize_t)>>('_PyOS_URandom');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<ffi.Void>, Py_ssize_t)>>(
+      '_PyOS_URandom');
   late final __PyOS_URandom =
       __PyOS_URandomPtr.asFunction<int Function(ffi.Pointer<ffi.Void>, int)>();
 
@@ -18309,9 +18310,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyOS_URandomNonblockPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<ffi.Void>, Py_ssize_t)>>('_PyOS_URandomNonblock');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<ffi.Void>, Py_ssize_t)>>(
+      '_PyOS_URandomNonblock');
   late final __PyOS_URandomNonblock = __PyOS_URandomNonblockPtr
       .asFunction<int Function(ffi.Pointer<ffi.Void>, int)>();
 
@@ -19722,9 +19723,9 @@ base class DartPythonCBindings {
   }
 
   late final __PyImport_IsInitializedPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<PyInterpreterState>)>>('_PyImport_IsInitialized');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyInterpreterState>)>>(
+      '_PyImport_IsInitialized');
   late final __PyImport_IsInitialized = __PyImport_IsInitializedPtr
       .asFunction<int Function(ffi.Pointer<PyInterpreterState>)>();
 
@@ -21192,9 +21193,9 @@ base class DartPythonCBindings {
   }
 
   late final _PySequence_DelItemPtr = _lookup<
-      ffi.NativeFunction<
-          ffi.Int Function(
-              ffi.Pointer<PyObject>, Py_ssize_t)>>('PySequence_DelItem');
+          ffi
+          .NativeFunction<ffi.Int Function(ffi.Pointer<PyObject>, Py_ssize_t)>>(
+      'PySequence_DelItem');
   late final _PySequence_DelItem = _PySequence_DelItemPtr.asFunction<
       int Function(ffi.Pointer<PyObject>, int)>();
 
@@ -22607,8 +22608,8 @@ final class PyAsyncMethods extends ffi.Struct {
 }
 
 typedef unaryfunc = ffi.Pointer<
-    ffi.NativeFunction<
-        ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
+    ffi
+    .NativeFunction<ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
 typedef sendfunc = ffi.Pointer<
     ffi.NativeFunction<
         ffi.Int32 Function(
@@ -22623,8 +22624,8 @@ abstract class PySendResult {
 }
 
 typedef reprfunc = ffi.Pointer<
-    ffi.NativeFunction<
-        ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
+    ffi
+    .NativeFunction<ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
 
 final class PyNumberMethods extends ffi.Struct {
   external binaryfunc nb_add;
@@ -22824,11 +22825,11 @@ typedef richcmpfunc = ffi.Pointer<
         ffi.Pointer<PyObject1> Function(
             ffi.Pointer<PyObject1>, ffi.Pointer<PyObject1>, ffi.Int)>>;
 typedef getiterfunc = ffi.Pointer<
-    ffi.NativeFunction<
-        ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
+    ffi
+    .NativeFunction<ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
 typedef iternextfunc = ffi.Pointer<
-    ffi.NativeFunction<
-        ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
+    ffi
+    .NativeFunction<ffi.Pointer<PyObject1> Function(ffi.Pointer<PyObject1>)>>;
 
 final class PyMethodDef extends ffi.Struct {
   external ffi.Pointer<ffi.Char> ml_name;
@@ -23057,7 +23058,7 @@ final class _ts extends ffi.Struct {
   external ffi.Pointer<PyObject> trash_delete_later;
 
   external ffi
-          .Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
+      .Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
       on_delete;
 
   external ffi.Pointer<ffi.Void> on_delete_data;
@@ -23252,7 +23253,7 @@ final class __sFILE extends ffi.Struct {
   external ffi.Pointer<ffi.Void> _cookie;
 
   external ffi
-          .Pointer<ffi.NativeFunction<ffi.Int Function(ffi.Pointer<ffi.Void>)>>
+      .Pointer<ffi.NativeFunction<ffi.Int Function(ffi.Pointer<ffi.Void>)>>
       _close;
 
   external ffi.Pointer<
@@ -23995,7 +23996,7 @@ final class _xid extends ffi.Struct {
               ffi.Pointer<_PyCrossInterpreterData>)>> new_object;
 
   external ffi
-          .Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
+      .Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Pointer<ffi.Void>)>>
       free;
 }