+++ packages/python_ffi_interface/CHANGELOG.md
@@ -1,10 +1,11 @@
-## NEXT
+## 0.2.0
 
 - Refactored to make use of Dart 3.x record types.
 - Utilized Python builtin function `iter` to convert Python iterables to Dart iterables.
 - Fixed
   [Bundling Python modules with external package](https://github.com/IVLIVS-III/dart_python_ffi/issues/10)
   .
+- Updated dependencies.
 
 ## 0.1.1